Refactor device registry unit tests, so that they can be used against multiple implementations
Move abstract tests to service-base

Tidying up and refactoring registration test to use proper interface

Moving all specific test related info to the concrete test implementations
