+++ services/device-registry/src/test/java/org/eclipse/hono/deviceregistry/FileBasedCredentialsServiceTest.java
@@ -13,8 +13,6 @@
 
 package org.eclipse.hono.deviceregistry;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doAnswer;
@@ -25,12 +23,15 @@ import static org.mockito.Mockito.when;
 import java.net.HttpURLConnection;
 import java.nio.charset.StandardCharsets;
 
+import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.eclipse.hono.auth.HonoPasswordEncoder;
-import org.eclipse.hono.service.credentials.CompleteCredentialsService;
+import org.eclipse.hono.service.credentials.AbstractCompleteCredentialsServiceTest;
+import org.eclipse.hono.service.credentials.CompleteBaseCredentialsService;
 import org.eclipse.hono.util.Constants;
 import org.eclipse.hono.util.CredentialsConstants;
 import org.eclipse.hono.util.CredentialsObject;
 import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.Timeout;
 import org.junit.runner.RunWith;
@@ -43,11 +44,9 @@ import io.vertx.core.Vertx;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.eventbus.EventBus;
 import io.vertx.core.file.FileSystem;
-import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
-import io.vertx.ext.unit.junit.VertxUnitRunner;
 
 
 /**
@@ -55,11 +54,12 @@ import io.vertx.ext.unit.junit.VertxUnitRunner;
  *
  */
 @RunWith(VertxUnitRunner.class)
-public class FileBasedCredentialsServiceTest {
+public class FileBasedCredentialsServiceTest extends AbstractCompleteCredentialsServiceTest {
 
     /**
      * Time out each test case after 5 seconds.
      */
+    @Rule
     public Timeout timeout = Timeout.seconds(5);
 
     private static final String FILE_NAME = "/credentials.json";
@@ -86,6 +86,11 @@ public class FileBasedCredentialsServiceTest {
         svc = new FileBasedCredentialsService(mock(HonoPasswordEncoder.class));
         svc.setConfig(props);
         svc.init(vertx, ctx);
+    }
+
+    @Override
+    public CompleteBaseCredentialsService getCompleteCredentialsService() {
+        return svc;
     }
 
     /**
@@ -296,154 +301,6 @@ public class FileBasedCredentialsServiceTest {
     }
 
     /**
-     * Verifies that only one set of credentials can be registered for an auth-id and type (per tenant).
-     *
-     * @param ctx The vert.x test context.
-     */
-    @Test
-    public void testAddCredentialsRefusesDuplicateRegistration(final TestContext ctx) {
-
-        register(svc, "tenant", "device", "myId", "myType", ctx);
-
-        final JsonObject payload2 = new JsonObject()
-                .put(CredentialsConstants.FIELD_PAYLOAD_DEVICE_ID, "other-device")
-                .put(CredentialsConstants.FIELD_AUTH_ID, "myId")
-                .put(CredentialsConstants.FIELD_TYPE, "myType")
-                .put(CredentialsConstants.FIELD_SECRETS, new JsonArray());
-        final Async add = ctx.async();
-        svc.add("tenant", payload2, ctx.asyncAssertSuccess(s -> {
-            assertThat(s.getStatus(), is(HttpURLConnection.HTTP_CONFLICT));
-            add.complete();
-        }));
-        add.await();
-    }
-
-    /**
-     * Verifies that the service returns 404 if a client wants to retrieve non-existing credentials.
-     *
-     * @param ctx The vert.x test context.
-     */
-    @Test
-    public void testGetCredentialsFailsForNonExistingCredentials(final TestContext ctx) {
-
-        final Async get = ctx.async();
-        svc.get("tenant", "myType", "non-existing", ctx.asyncAssertSuccess(s -> {
-                    assertThat(s.getStatus(), is(HttpURLConnection.HTTP_NOT_FOUND));
-                    get.complete();
-                }));
-        get.await();
-    }
-
-    /**
-     * Verifies that the service returns existing credentials.
-     *
-     * @param ctx The vert.x test context.
-     */
-    @Test
-    public void testGetCredentialsSucceedsForExistingCredentials(final TestContext ctx) {
-
-        register(svc, "tenant", "device", "myId", "myType", ctx);
-
-        final Async get = ctx.async();
-        svc.get("tenant", "myType", "myId", ctx.asyncAssertSuccess(s -> {
-                    assertThat(s.getStatus(), is(HttpURLConnection.HTTP_OK));
-                    assertThat(s.getPayload().getString(CredentialsConstants.FIELD_AUTH_ID), is("myId"));
-                    assertThat(s.getPayload().getString(CredentialsConstants.FIELD_TYPE), is("myType"));
-                    get.complete();
-                }));
-        get.await();
-    }
-
-    /**
-     * Verifies that service returns existing credentials for proper client context.
-     *
-     * @param ctx The vert.x test context.
-     */
-    @Test
-    public void testGetCredentialsSucceedsForProperClientContext(final TestContext ctx) {
-        final JsonObject clientContext = new JsonObject()
-            .put("client-id", "gateway-one");
-
-        register(svc, "tenant", "device", "myId", "myType", clientContext, new JsonArray(), ctx);
-
-
-        final Async get = ctx.async();
-        svc.get("tenant", "myType", "myId", clientContext, ctx.asyncAssertSuccess(s -> {
-            assertThat(s.getStatus(), is(HttpURLConnection.HTTP_OK));
-            assertThat(s.getPayload().getString(CredentialsConstants.FIELD_AUTH_ID), is("myId"));
-            assertThat(s.getPayload().getString(CredentialsConstants.FIELD_TYPE), is("myType"));
-            get.complete();
-        }));
-        get.await(2000);
-    }
-
-    /**
-     * Verifies that service returns 404 if a client provides wrong client context.
-     *
-     * @param ctx The vert.x test context.
-     */
-    @Test
-    public void testGetCredentialsFailsForWrongClientContext(final TestContext ctx) {
-        final JsonObject clientContext = new JsonObject()
-            .put("client-id", "gateway-one");
-
-        register(svc, "tenant", "device", "myId", "myType", clientContext, new JsonArray(), ctx);
-
-        final JsonObject testContext = new JsonObject()
-            .put("client-id", "gateway-two");
-
-        final Async get = ctx.async();
-        svc.get("tenant", "myType", "myId", testContext, ctx.asyncAssertSuccess(s -> {
-            assertThat(s.getStatus(), is(HttpURLConnection.HTTP_NOT_FOUND));
-            get.complete();
-        }));
-        get.await(2000);
-    }
-
-    /**
-     * Verifies that the service removes credentials for a given auth-id and type.
-     *
-     * @param ctx The vert.x test context.
-     */
-    @Test
-    public void testRemoveCredentialsByAuthIdAndTypeSucceeds(final TestContext ctx) {
-
-        register(svc, "tenant", "device", "myId", "myType", ctx);
-
-        final Async remove = ctx.async();
-        svc.remove("tenant", "myType", "myId", ctx.asyncAssertSuccess(s -> {
-            assertThat(s.getStatus(), is(HttpURLConnection.HTTP_NO_CONTENT));
-            assertNotRegistered(svc, "tenant", "myId", "myType", ctx);
-            remove.complete();
-        }));
-        remove.await();
-    }
-
-    /**
-     * Verifies that the service removes all credentials for a device but keeps credentials
-     * of other devices.
-     *
-     * @param ctx The vert.x test context.
-     */
-    @Test
-    public void testRemoveCredentialsByDeviceSucceeds(final TestContext ctx) {
-
-        register(svc, "tenant", "device", "myId", "myType", ctx);
-        register(svc, "tenant", "device", "myOtherId", "myOtherType", ctx);
-        register(svc, "tenant", "other-device", "thirdId", "myType", ctx);
-
-        final Async remove = ctx.async();
-        svc.removeAll("tenant", "device", ctx.asyncAssertSuccess(s -> {
-            assertThat(s.getStatus(), is(HttpURLConnection.HTTP_NO_CONTENT));
-            assertNotRegistered(svc, "tenant", "myId", "myType", ctx);
-            assertNotRegistered(svc, "tenant", "myOtherId", "myOtherType", ctx);
-            assertRegistered(svc, "tenant", "thirdId", "myType", ctx);
-            remove.complete();
-        }));
-        remove.await();
-    }
-
-    /**
      * Verifies that the <em>modificationEnabled</em> property prevents updating an existing entry.
      *
      * @param ctx The vert.x test context.
@@ -454,7 +311,7 @@ public class FileBasedCredentialsServiceTest {
         // GIVEN a registry containing a set of credentials
         // that has been configured to not allow modification of entries
         props.setModificationEnabled(false);
-        register(svc, "tenant", "device", "myId", "myType", ctx);
+        register(getCompleteCredentialsService(), "tenant", "device", "myId", "myType", ctx);
 
         // WHEN trying to update the credentials
         final Async updateFailure = ctx.async();
@@ -478,7 +335,7 @@ public class FileBasedCredentialsServiceTest {
         // GIVEN a registry containing a set of credentials
         // that has been configured to not allow modification of entries
         props.setModificationEnabled(false);
-        register(svc, "tenant", "device", "myId", "myType", ctx);
+        register(getCompleteCredentialsService(), "tenant", "device", "myId", "myType", ctx);
 
         // WHEN trying to remove the credentials
         final Async removeFailure = ctx.async();
@@ -489,73 +346,5 @@ public class FileBasedCredentialsServiceTest {
 
         // THEN the removal fails
         removeFailure.await();
-    }
-
-    private static void assertRegistered(
-            final CompleteCredentialsService svc,
-            final String tenant,
-            final String authId,
-            final String type,
-            final TestContext ctx) {
-
-        final Async registration = ctx.async();
-        svc.get(tenant, type, authId, ctx.asyncAssertSuccess(t -> {
-            assertThat(t.getStatus(), is(HttpURLConnection.HTTP_OK));
-            registration.complete();
-        }));
-        registration.await();
-    }
-
-    private static void assertNotRegistered(
-            final CompleteCredentialsService svc,
-            final String tenant,
-            final String authId,
-            final String type,
-            final TestContext ctx) {
-
-        final Async registration = ctx.async();
-        svc.get(tenant, type, authId, ctx.asyncAssertSuccess(t -> {
-            assertThat(t.getStatus(), is(HttpURLConnection.HTTP_NOT_FOUND));
-            registration.complete();
-        }));
-        registration.await();
-    }
-
-    private static void register(
-            final CompleteCredentialsService svc,
-            final String tenant,
-            final String deviceId,
-            final String authId,
-            final String type,
-            final TestContext ctx) {
-        register(svc, tenant, deviceId, authId, type, null, new JsonArray(), ctx);
-    }
-
-    private static void register(
-            final CompleteCredentialsService svc,
-            final String tenant,
-            final String deviceId,
-            final String authId,
-            final String type,
-            final JsonObject clientContext,
-            final JsonArray secrets,
-            final TestContext ctx) {
-
-        final JsonObject data = new JsonObject()
-                .put(CredentialsConstants.FIELD_PAYLOAD_DEVICE_ID, deviceId)
-                .put(CredentialsConstants.FIELD_AUTH_ID, authId)
-                .put(CredentialsConstants.FIELD_TYPE, type)
-                .put(CredentialsConstants.FIELD_SECRETS, secrets);
-
-        if (clientContext != null) {
-            data.mergeIn(clientContext);
-        }
-
-        final Async registration = ctx.async();
-        svc.add("tenant", data, ctx.asyncAssertSuccess(s -> {
-            assertThat(s.getStatus(), is(HttpURLConnection.HTTP_CREATED));
-            registration.complete();
-        }));
-        registration.await();
     }
 }
